Add ReportTemplate for term members
Allow a different template for each type of report, and add an initial
basic one for Term memberships.
