+++ lib/wikidata_position_history/report/term.rb
@@ -8,12 +8,25 @@ module WikidataPositionHistory
         SPARQL::TermMandatesQuery
       end
 
+      def template_class
+        ReportTemplate::Term
+      end
+
       def mandate_class
         TermMandateRow
       end
 
       def biodata_query
         SPARQL::TermBioQuery
+      end
+
+      def table_rows
+        mandates.map do |mandate|
+          {
+            mandate: OutputRow::TermMandate.new(mandate),
+            bio:     biodata_for(mandate.officeholder),
+          }
+        end
       end
     end
   end
